Char sequence token
